+++ hardpy/pytest_hardpy/db/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2024 Everypin
+# Copyright (c) 2025 Everypin
 # GNU General Public License v3.0 (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from hardpy.pytest_hardpy.db.base_store import BaseStore
@@ -13,6 +13,7 @@ from hardpy.pytest_hardpy.db.stand_type import (
     SubUnit,
 )
 from hardpy.pytest_hardpy.db.statestore import StateStore
+from hardpy.pytest_hardpy.db.tempstore import TempStore
 
 __all__ = [
     "BaseStore",
@@ -26,4 +27,5 @@ __all__ = [
     "StateStore",
     "StringMeasurement",
     "SubUnit",
+    "TempStore",
 ]