+++ hardpy/pytest_hardpy/pytest_wrapper.py
@@ -89,6 +89,8 @@ class PyTestWrapper:
 
         if self.config.stand_cloud.connection_only:
             cmd.append("--sc-connection-only")
+        if self.config.stand_cloud.autosync:
+            cmd.append("--sc-autosync")
         cmd.append("--hardpy-pt")
         if start_args:
             for key, value in start_args.items():
@@ -101,7 +103,7 @@ class PyTestWrapper:
                 cwd=self._config_manager.tests_path,
                 creationflags=subprocess.CREATE_NEW_PROCESS_GROUP,
             )
-        if system() == "Linux":
+        else:
             self._proc = subprocess.Popen(  # noqa: S603
                 cmd,
                 cwd=self._config_manager.tests_path,
@@ -116,10 +118,10 @@ class PyTestWrapper:
             bool: True if pytest was running and stopped
         """
         if self.is_running() and self._proc:
-            if system() == "Linux":
-                self._proc.terminate()
-            elif system() == "Windows":
+            if system() == "Windows":
                 self._proc.send_signal(signal.CTRL_BREAK_EVENT)  # type: ignore
+            else:
+                self._proc.terminate()
             return True
         return False